unified the network and dispatcher crates; renamed the dispatcher to bridge and separated the setup function from it
