--- crates/network/dispatcher/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "dispatcher"
-version = "0.1.0"
-edition = "2024"
-description = "Crypt of the Lost network event dispatcher"
-license-file = "../../LICENSE"
-repository = "https://github.com/Sietse2202/crypts-of-the-lost"
-readme = "../../README.md"
-keywords = ["network", "bevy-plugin", "bevy"]
-categories = ["games"]
-
-[dependencies]
-protocol.workspace = true
-bevy.workspace = true
-tokio.workspace = true
-tracing.workspace = true
-handler.workspace = true
-config.workspace = true
-
-[lints]
-workspace = true
