+++ crates/network/src/handler/serialize.rs
